JWT Integration - Code Cleanup & LHS Check
